+++ packages/vda-node-client/src/getDataCenters.ts
@@ -1,39 +1,18 @@
-import { CONTRACT_ADDRESS, CONTRACT_ABI as abiList, RPC_URLS } from "@verida/vda-common";
-
-import { JsonRpcProvider } from '@ethersproject/providers';
-import { BigNumberish, Contract } from 'ethers';
-
+import { BigNumberish } from "ethers";
+import { executeFunction } from "./utils";
+import { EnumStatus } from "@verida/types";
 
 /**
  * Call getDatacenters() function of `StorageNodeRegistry` contract
  * @param ids Array of datacenterIds
  */
-export async function isDataCenterNameRegistered(network: string, name: string) {
-    const rpcUrl = RPC_URLS[network]
-    if (!rpcUrl) {
-        throw new Error(`Unable to locate RPC_URL for network: ${network}`)
-    }
-
-    // Simple read-only of the blockchain
-
-    const contractABI = abiList["StorageNodeRegistry"];
-    const provider = new JsonRpcProvider(rpcUrl);
-    const address = CONTRACT_ADDRESS["StorageNodeRegistry"][network];
-
-    if (!address) {
-        throw new Error(`Empty contract address for network-${network}`)
-    }
-
-    const contract = new Contract(address, contractABI.abi, provider);
-    
-    let data;
-    try {
-        data = (await contract.callStatic.isDataCenterNameRegistered(name));
-    } catch (e: any) {
-        throw new Error('Failed to check datacenter name');
-    }
-
-    return data
+export async function isRegisteredDataCenterName(network: string, name: string) {
+    return await executeFunction(
+        network,
+        'isRegisteredDataCenterName',
+        'Failed to check datacenter name',
+        name
+    );
 }
   
 /**
@@ -41,32 +20,12 @@ export async function isDataCenterNameRegistered(network: string, name: string) 
  * @param ids Array of datacenterIds
  */
 export async function getDataCenters(network: string, ids: BigNumberish[]) {
-    const rpcUrl = RPC_URLS[network]
-    if (!rpcUrl) {
-        throw new Error(`Unable to locate RPC_URL for network: ${network}`)
-    }
-
-    // Simple read-only of the blockchain
-
-    const contractABI = abiList["StorageNodeRegistry"];
-    const provider = new JsonRpcProvider(rpcUrl);
-    const address = CONTRACT_ADDRESS["StorageNodeRegistry"][network];
-
-    if (!address) {
-        throw new Error(`Empty contract address for network-${network}`)
-    }
-
-    const contract = new Contract(address, contractABI.abi, provider);
-    
-    let data;
-    try {
-        data = (await contract.callStatic.getDataCenters(ids));
-    } catch (err: any) {
-        const message = err.reason ? err.reason : err.message;
-        throw new Error(`Failed to get datacenters (${message})`);
-    }
-
-    return data
+    return await executeFunction(
+        network,
+        'getDataCenters',
+        'Failed to get datacenters by id',
+        ids
+    );
 }
 
 /**
@@ -74,96 +33,66 @@ export async function getDataCenters(network: string, ids: BigNumberish[]) {
  * @param names Array of name
  */
 export async function getDataCentersByName(network: string, names: string[]) {
-    const rpcUrl = RPC_URLS[network]
-    if (!rpcUrl) {
-        throw new Error(`Unable to locate RPC_URL for network: ${network}`)
-    }
-
-    // Simple read-only of the blockchain
-
-    const contractABI = abiList["StorageNodeRegistry"];
-    const provider = new JsonRpcProvider(rpcUrl);
-    const address = CONTRACT_ADDRESS["StorageNodeRegistry"][network];
-
-    if (!address) {
-        throw new Error(`Empty contract address for network-${network}`)
-    }
-
-    const contract = new Contract(address, contractABI.abi, provider);
-    
-    let data;
-    try {
-        data = (await contract.callStatic.getDataCentersByName(names));
-    } catch (err: any) {
-        const message = err.reason ? err.reason : err.message;
-        throw new Error(`Failed to get datacenters by names (${message})`);
-    }
-
-    return data
+    return await executeFunction(
+        network,
+        'getDataCentersByName',
+        'Failed to get datacenters by name',
+        names
+    );
 }
 
 /**
  * Call getDatacentersByCountry() function of `StorageNodeRegistry` contract
- * @param ids Array of datacenterIds
+ * @param network Target network
+ * @param countryCode Country code of data centers
+ * @param status Status of data centers
  */
-export async function getDataCentersByCountry(network: string, countryCode: string) {
-    const rpcUrl = RPC_URLS[network]
-    if (!rpcUrl) {
-        throw new Error(`Unable to locate RPC_URL for network: ${network}`)
+export async function getDataCentersByCountry(
+    network: string, 
+    countryCode: string, 
+    status?: EnumStatus) {
+    if (status === undefined) {
+        return await executeFunction(
+            network,
+            'getDataCentersByCountry',
+            'Failed to get datacenters by country',
+            countryCode
+        );
+    } else {
+        return await executeFunction(
+            network,
+            'getDataCentersByCountryAndStatus',
+            'Failed to get datacenters by country and status',
+            countryCode,
+            status
+        );
     }
-
-    // Simple read-only of the blockchain
-
-    const contractABI = abiList["StorageNodeRegistry"];
-    const provider = new JsonRpcProvider(rpcUrl);
-    const address = CONTRACT_ADDRESS["StorageNodeRegistry"][network];
-
-    if (!address) {
-        throw new Error(`Empty contract address for network-${network}`)
-    }
-
-    const contract = new Contract(address, contractABI.abi, provider);
-    
-    let data;
-    try {
-        data = (await contract.callStatic.getDataCentersByCountry(countryCode));
-    } catch (err: any) {
-        const message = err.reason ? err.reason : err.message;
-        throw new Error(`Failed to get datacenters by country (${message})`);
-    }
-
-    return data
 }
 
 /**
  * Call getDatacentersByRegion() function of `StorageNodeRegistry` contract
- * @param ids Array of datacenterIds
+ * @param network Target network
+ * @param regionCode Region code of data centers
+ * @param status Status of data centers
  */
-export async function getDataCentersByRegion(network: string, regionCode: string) {
-    const rpcUrl = RPC_URLS[network]
-    if (!rpcUrl) {
-        throw new Error(`Unable to locate RPC_URL for network: ${network}`)
+export async function getDataCentersByRegion(
+    network: string, 
+    regionCode: string, 
+    status?: EnumStatus) {
+    if (status === undefined) {
+        return await executeFunction(
+            network,
+            'getDataCentersByRegion',
+            'Failed to get datacenters by region',
+            regionCode
+        );
+    } else {
+        return await executeFunction(
+            network,
+            'getDataCentersByRegionAndStatus',
+            'Failed to get datacenters by region and status',
+            regionCode,
+            status
+        );
     }
-
-    // Simple read-only of the blockchain
-
-    const contractABI = abiList["StorageNodeRegistry"];
-    const provider = new JsonRpcProvider(rpcUrl);
-    const address = CONTRACT_ADDRESS["StorageNodeRegistry"][network];
-
-    if (!address) {
-        throw new Error(`Empty contract address for network-${network}`)
-    }
-
-    const contract = new Contract(address, contractABI.abi, provider);
-    
-    let data;
-    try {
-        data = (await contract.callStatic.getDataCentersByRegion(regionCode));
-    } catch (err: any) {
-        const message = err.reason ? err.reason : err.message;
-        throw new Error(`Failed to get datacenters by region (${message})`);
-    }
-
-    return data
 }