+++ packages/vda-node-client/src/getVDATokenAddress.ts
@@ -3,10 +3,10 @@ import { executeFunction } from "./utils";
 /**
  * Call DECIMAL() function of `StorageNodeRegistry` contract
  */
-export async function getContractDecimal(network: string): Promise<number> {
+export async function getVDATokenAddress(network: string): Promise<string> {
     return await executeFunction(
         network,
-        'DECIMAL',
-        'Failed to get DECIMAL'
+        'getVDATokenAddress',
+        'Failed to get Verida Token Address'
     );
 }