+++ packages/vda-node-client/test/getNodes.test.ts
@@ -1,18 +1,67 @@
 const assert = require("assert");
 import { BigNumber, Wallet } from 'ethers';
-import { getNodeByAddress, getNodeByEndpoint, getNodesByCountry, getNodesByRegion } from '../src/getNodes';
-import { REGISTERED_DIDS, DID_NODE_MAP } from '@verida/vda-common-test';
+import { getNodeByAddress, getNodeByEndpoint, getNodeByName, getNodesByCountry, getNodesByRegion, isRegisteredNodeAddress, isRegisteredNodeEndpoint, isRegisteredNodeName } from '../src/getNodes';
+import { REGISTERED_DIDS, DID_NODE_MAP, REMOVE_START_DIDS } from '@verida/vda-common-test';
+import { EnumStatus } from '@verida/types';
 
 const NETWORK = 'testnet';
 
 describe('getNodes test', function() {
     this.timeout(20000);
 
-    it('getNodeByAddress',async () => {
-        // Return `undefined` for unregistered DID address
-        let result = await getNodeByAddress(NETWORK, Wallet.createRandom().address);
-        assert.equal(result, undefined, `Return 'undefined' for unregistered DID address`);
+    let registeredNodes: any[] = [];
+    before(() => {
+        DID_NODE_MAP.forEach((value) => registeredNodes.push(value))
+    })
 
+    it('isRegisteredNodeName',async () => {
+        let result = await isRegisteredNodeName(NETWORK, "InvalidName");
+        assert.ok(result === false, 'false for unregistered name');
+
+        result = await isRegisteredNodeName(NETWORK, registeredNodes[0].name);
+        assert.ok(result === true, 'true for registered name');
+    })
+
+    it('isRegisteredNodeAddress',async () => {
+        let result = await isRegisteredNodeAddress(NETWORK, Wallet.createRandom().address);
+        assert.ok(result === false, 'false for unregistered address');
+
+        const wallet = new Wallet(REGISTERED_DIDS[0].privateKey);
+        result = await isRegisteredNodeAddress(NETWORK, wallet.address);
+        assert.ok(result === true, 'true for registered address');
+    })
+
+    it('isRegisteredNodeEndpoint',async () => {
+        let result = await isRegisteredNodeEndpoint(NETWORK, "https://invalid-uri");
+        assert.ok(result === false, 'false for unregistered uri');
+
+        result = await isRegisteredNodeEndpoint(NETWORK, registeredNodes[0].endpointUri);
+        assert.ok(result === true, 'true for registered uri');
+    })
+
+    it('getNodeByName',async () => {
+        // Rejected for unregistered name
+        let result
+        try {
+            result = await getNodeByName(NETWORK, 'InvalidName');
+        } catch (e) {
+            assert.ok(true, 'Rejected for unregistered name');
+        }
+        
+        // Return a node for registered name
+        result = await getNodeByName(NETWORK, registeredNodes[0].name);
+        assert.equal(typeof(result), 'object', 'Returned a node for registered DID address');
+    })
+
+    it('getNodeByAddress',async () => {
+        // Rejected for unregistered DID address
+        let result
+        try {
+            result = await getNodeByAddress(NETWORK, Wallet.createRandom().address);
+        } catch (e) {
+            assert.ok(true, 'Rejected for unregistered address');
+        }
+        
         // Return a node for registered DID address
         const did = new Wallet(REGISTERED_DIDS[0].privateKey);
         result = await getNodeByAddress(NETWORK, did.address);
@@ -20,13 +69,17 @@ describe('getNodes test', function() {
     })
 
     it('getNodeByEndpoint',async () => {
-        // Return `undefined` for unregistered endpointURI
-        let result = await getNodeByEndpoint(NETWORK, `https://${Wallet.createRandom().address}`);
-        assert.equal(result, undefined, `Return 'undefined' for unregistered endpointURI`);
-
-        const registeredNode = DID_NODE_MAP.get(REGISTERED_DIDS[0].address);
-        result = await getNodeByEndpoint(NETWORK, registeredNode.endpointUri);
-        assert.equal(typeof(result), 'object', 'Returned a node for registered endpointURI');
+        // Rejected for unregistered endpoint URI
+        let result
+        try {
+            result = await getNodeByEndpoint(NETWORK, `https://${Wallet.createRandom().address}`);
+        } catch (e) {
+            assert.ok(true, 'Rejected for unregistered address');
+        }
+        
+        // Return a node for registered DID address
+        result = await getNodeByEndpoint(NETWORK, registeredNodes[0].endpointUri);
+        assert.equal(typeof(result), 'object', 'Returned a node for registered DID address');
     })
 
     it('getNodesByCountry',async () => {
@@ -42,6 +95,12 @@ describe('getNodes test', function() {
         }
     })
 
+    it('getNodesByCountryAndStatus',async () => {
+        const removedCountryCode = (DID_NODE_MAP.get(REMOVE_START_DIDS[0].address)).countryCode;
+        const result = await getNodesByCountry(NETWORK, removedCountryCode, EnumStatus.removing);
+        assert.ok(result.length > 0, "Return 1 or more removing nodes for registered country code");
+    })
+
     it('getNodesByRegion',async () => {
         const regionList: any[] = [];
         DID_NODE_MAP.forEach((node) => {
@@ -54,4 +113,11 @@ describe('getNodes test', function() {
             assert.ok(result.length > 0, "Return 1 or more nodes for registered region code");
         }
     })
+
+    it('getNodesByCountryAndStatus',async () => {
+        const removedRegioinCode = (DID_NODE_MAP.get(REMOVE_START_DIDS[0].address)).regionCode;
+        const result = await getNodesByRegion(NETWORK, removedRegioinCode, EnumStatus.removing);
+        assert.ok(result.length > 0, "Return 1 or more removing nodes for registered country code");
+    })
+
 })